change component according atomic designe, rename stories, remove scss
